requirements.txt
SQLAlchemy>=2.0,<3.0
PyMySQL>=1.1,<2.0
python-dotenv>=1.0,<2.0
requests>=2.31,<3.0
tenacity>=8.2,<9.0
pytest>=8.3.4
pytest-cov>=6.0.0
responses>=0.25.3